This version has removed dead functionality (e.g. move pool to unpools and stripped out all shm, ttl, gc related code) to clean compile.  I have sorted out the interfaces statically (albeit with most implementation content commented out), so that I can get to clean but non-functional compile.  This includes remapping of some redundant interfaces to allow the implementation source to remain (largely) unchanged.  Note that this version isn’t intended to be runnable.
Infact php -i works fine, which is more than I expected.

This commit establishes a new baseline, and the next stage is to sort out and debug the cache
implementation so that LPC works in its culled APC mode: that is with a full "shared" cache
but in unpool process space with the old compile to shared cache mode.