== LPC ==  Another working version of APC-lite functionality and a commit way-point before the next round of changes.  This version:     * Dumps the stack handling -- I only use ToS in LPC so the stacking is redundant     * Extend the function entry collection (routines are indented by the stack depth     * Adds my "swizzling almost free" algos, plus (temporary validation) logic.
The relocation concept seems to be OK in that I am picking up all pointers to be relocated
and no memory leaks.  I need to extend the pool handle to support relocation and gut the
cache to swap in a CacheDB-based cache.  But time for another waypoint commit before I
start this work.