== LPC ==  A clean compile of APC-lite functionality and a part implementation of pools to support relocation (I still have to code up the load function).
I've also dumped the APC REGEX wrapper, becuase this is mainly to support (i) REGEX caching which
PHP regexp engine goes to great lengths to do already; (ii) to support PHP versions < 5.2.2 and I
simply drop REGEX in this case; (iii) these array and negative assertions are standard REGEX
functionality anyway, so again why bother.

Really, the only reason for this commit is that I am a bit twitchy about the volume of code changes
since the last change.  The next waypoint commit will be a working version of all this.