== LPC ==  Another clean compile of APC-lite functionality and implementation of pools to support relocation, including the load function, and begugging aftre unit test shake-down.
In this version the lpc serializer code which was only used for user variable caching has also been
dropped.

I have decided to rationalise the pool implimentation, as I understand the internals of the compile
process more.  We need two type of pool: a (PICable) serial pool for moving to external storage and
a executable pool for use by the Zend engine, with the corresponding types LPC_SERIALPOOL and
LPC_EXECPOOL.  This is going to trigger another round of refactoring, so I've decided to make a
waypoint commit, before I start this given the volume of code changes since the last change