=== LPC === The major rework arising from the 14-16 Dec review has now largely been implemented and is working: excluding memory leak reports, only 4 of the tests fail and all of these with subtle variations on error path handling.  However, this leakage is still a major issue and since the intro of interning in the serial pools, this has grown from a small trickle in a few tests to some fairly large reports (e.g test47 has 89 leaks reported).
I want to do a full review of de-interning reuse, but this way-point establishes a baseline for any
changes should I decide to back them out.

Note that TECHNOTES.txt and TODO are up to date so comparing the current and previous versions
give a more detailed overview of this last set of changes.