=== LPC ===
Interning/de-interning is working well -- at least at Zend 2.3; Zend 2.4 comes later

I've done a review of the zend_compiler's treatment of class inheritence and havve decided that
the current processing -- which largley follows the APC approach -- needs further refactoring.
The nub of this issue is that the Zend compiler by default defers the inheritance binding of a
child class to its parent to runtime and implements this through the ZEND_DECLARE_INHERITED_CLASS
zend_op.  However, if the parent class is already register at compile-time then the
zend_do_inheritance() is executed as part of the compile and the ZEND_FETCH_CLASS /
ZEND_DECLARE_INHERITED_CLASS zend_op pair are NOPped out.  However, this strategy isn't viable
for LPC, so such inheritance binding needs to be backed out, and re-executed on cache load.

Given the coming round of refacting, this way-point establishes a baseline for any changes
should I decide to back them out. Note that TECHNOTES.txt and TODO are up to date so comparing
the current and previous versions give a more detailed overview of this last set of changes.