LPC:  freezing development following Zend's decision to open-source OptimizerPlus
When I started the development of LPC it was a proof of concept to demonstrate
that a file-based opcode cache could deliver performance improvements.  As I
discussed in TECHNOTES.txt, my initial aim was to triage the APC code, and adopt a
minimum chagne policy.  Unfortunately because of features of the APC implementation,
I found this unachievabe and ended up doing a rewrite.  This allowed me to make
significant progress but it was killer re production use as there would be not
desire on the part of the PHP community to accept yet another opcode cache.

O+ is more closely aligned to the sort of architecture that I want to work with
and so I have decided to regress much of the LPC technology into an O+ fork with
the aim of proving this concept in a format that might ultimately see adoption.

This commit is simply a checkpoint of my work-in-progress of LPC at this point.